Update old tests with new random module & refactoring (renaming) files
