+++ tests/Randomizer.h
@@ -1,45 +1,26 @@
-/*
-   Enki - a fast 2D robot simulator
-   Copyright © 2017 Nicolas Palard <nicolas.palard@etu.u-bordeaux.fr>
+#ifndef __RANDOMIZER_H
+#define __RANDOMIZER_H
 
-   This program is free software: you can redistribute it and/or modify
-   it under the terms of the GNU General Public License as published by
-   the Free Software Foundation, either version 2 of the License, or
-   (at your option) any later version.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty of
-   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-   GNU General Public License for more details.
-
-   You should have received a copy of the GNU General Public License
-   along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef RANDOM_WORLD_H
-#define RANDOM_WORLD_H
+#include <random>
+#include <chrono>
 
 #include <enki/PhysicalEngine.h>
 #include <enki/robots/thymio2/Thymio2.h>
 #include <enki/robots/e-puck/EPuck.h>
 #include <enki/robots/s-bot/Sbot.h>
-#include <enki/robots/marxbot/Marxbot.h>
 #include <enki/robots/khepera/Khepera.h>
+#include <enki/robots/marxbot/Marxbot.h>
 
-/*!	\file Randomizer.h
-	\brief Header of 'random' module.
+/*!
+	\file Randomizer.h
+	\brief This file provides a way for random world generation and
+	random object generation.
 */
 
 namespace Enki
 {
-	// Those enumeration items end with a '_' so that they don't collide with
-	// the ones used in the serialization module.
-	enum ROBOT_TYPES { THYMIO2_, EPUCK_, SBOT_, MARXBOT_, KHEPERA_ };
+	const int NUMBER_OF_ROBOTS_TYPES = 5;
 
-	const int MAX_ROBOTS = 30;
-	const bool ONLY_THYMIOS = true;
-	const int MAX_DYNAMIC_OBJECTS = 50;
-	const bool ONLY_SIMPLE_SHAPES = true;
 	const int MAX_HEIGHT = 800;
 	const int MIN_HEIGHT = 200;
 	const int MAX_WIDTH = 800;
@@ -47,109 +28,184 @@ namespace Enki
 	const int MIN_RADIUS = 300;
 	const int MAX_RADIUS = 500;
 
-	const int NUMBER_OF_ROBOTS_TYPES = 5;
+	class Randomizer
+	{
+	public:
+		enum ROBOT_TYPES {
+			THYMIO2_,
+			EPUCK_,
+			SBOT_,
+			MARXBOT_,
+			KHEPERA_,
+		};
 
-	/*!
-		\brief Return a random Integer between min and max.
-		\param min the minimum
-		\param max the maximum
-		\pre min should be not equals to max
-		\return The generated Integer
-	*/
-	int randomNumber(int min, int max);
+	private:
+		World* world;
+		long long int seed;
+		// We use Mersenne Twister 19937 engine because
+		// it's the most popular engine for
+		// PRNG (Pseudo random number generation)
+		// Mersenne Twister engine procude better output
+		// than the default engine but it is way slower
+		// /!\ This is not thread safe.
+		std::mt19937 randomEngine;
 
-	/*!
-		\brief Generate randomly a Color.
-		\post The returned color should be compound with 4 float values between 0 a 1.
-		\return The generated PhysicalObject
-	*/
-	Color randomColor();
+		std::uniform_real_distribution<> color_distr = std::uniform_real_distribution<>(0.0, 1.0);
+		std::uniform_int_distribution<> bool_distr = std::uniform_int_distribution<>(0, 1);
 
-	/*!
-		\brief Generate randomly a PhysicalObject
-		\param w the world in which the object will spawn.
-		\pre w must have been initialized
-		\post The PhysicalObject should have his x,y in the given World.
-		\return The generated PhysicalObject
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	PhysicalObject* randomObject(World* w);
+	public:
+		/*!
+			\brief Constructor of a Randomizer.
+			 A randomizer aims to generate random Enki Worlds and Enki
+			Objects.
+			\param w a pre-constructed world in which it will generate objects.
+		*/
+		Randomizer(World* w, long long int seed=-1);
 
-	/*!
-		\brief Generic function used to create randomly any kind of robots
-		\param w the world in which the robots will spawn
-		\pre w must have been initialized
-		\return The generated robot
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	Robot* randomRobot(World* w);
+		/*!
+			\brief Constructor of a Randomizer.
+			A randomizer aims to generate random Enki Worlds and Enki
+			Objects.
+			This constructor will generate a random world because no
+			base world is specified.
+		*/
+		Randomizer(long long int seed = -1);
 
-	/*!
-		\brief Generate randomly a Thymio2
-		\param w the world in which the thymio will spawn.
-		\pre w must have been initialized
-		\post The Thymio2 should have his x,y in the given World.
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	Thymio2* randomThymio(World* w);
+		~Randomizer();
 
-	/*!
-		\brief Generate randomly a EPuck
-		\param w the world in which the thymio will spawn.
-		\pre w must have been initialized
-		\post The EPuck should have his x,y in the given World.
-		\return The generated EPuck
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	EPuck* randomEPuck(World* w);
+		/*!
+			\brief This function allows debug by returning the seed
+			of the current random engine.
+			\return the seed of the current random engine.
+		*/
+		long long int getSeed();
 
-	/*!
-		\brief Generate randomly a Sbot
-		\param w the world in which the thymio will spawn.
-		\pre w must have been initialized
-		\post The Sbot should have his x,y in the given World.
-		\return The generated Sbot
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	Sbot* randomSbot(World* w);
+		/*!
+			\brief This functions allows to set a custom seed to
+			the current random engine.
+			\param seed the seed to set.
+		*/
+		void setSeed(const long long int &seed);
 
-	/*!
-		\brief Generate randomly a Marxbot
-		\param w the world in which the thymio will spawn.
-		\pre w must have been initialized
-		\post The Marxbot should have his x,y in the given World.
-		\return The generated Marxbot
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	Marxbot* randomMarxbot(World* w);
+		/*!
+			\brief This functions allows the user to work on a
+			random world outside of the randomizer module.
+			This can be used to tweak worlds.
+			\return the actual randomized world.
+		*/
+		World* getWorld();
 
-	/*!
-		\brief Generate randomly a Khepera
-		\param w the world in which the thymio will spawn.
-		\pre w must have been initialized
-		\post The Khepera should have his x,y in the given World.
-		\return The generated Khepera
-	*/
-	// TODO: Consider removing world parameter to gain flexibility
-	Khepera* randomKhepera(World* w);
+		/*!
+			\brief This function is used to start a new world generation.
+		*/
+		void resetWorld();
 
-	/*!
-		\brief Generate randomly an empty world
-		\return The generated World
-	*/
-	World* randomWorld();
+		/*!
+			\brief This function aims to create an empty randomized world.
+			It is used in the constructor when no world is gave.
+			It only creates the basis of a World : it's type and size.
+			\return the created world.
+		*/
+		World* randWorld();
 
-	/*!
-		\brief Add random generated objects to the world passed in parameter
-		\param w the world in which the objects are going to be added.
-	*/
-	void addObjects(World *w);
+		/*!
+			\brief This function generates any kind of enki objects
+			such as robots and physical objects.
+			\return the generated object.
+		*/
+		PhysicalObject* randObject();
 
-	/*!
-		\brief Add random generated robots to the world passed in parameter
-		\param w the world in which the robots are going to be added.
-	*/
-	void addRobots(World *w);
+		/*!
+			\brief This function generates a PhysicalObject that is not
+			a robot.
+			\param hullSize the hull size can be specified.
+			\return the generated object.
+		*/
+		PhysicalObject* randPhysicalObject(const int &hullSize = -1);
+
+		/*!
+			\brief This function generates a Robot.
+			It can generate: Thymio2, SBot, Epuck, Marxbot, Khepera
+			\return the generated robot.
+		*/
+		Robot* randRobot();
+		Thymio2* randThymio();
+		EPuck* randEPuck();
+		Khepera* randKhepera();
+		Sbot* randSbot();
+		Marxbot* randMarxbot();
+
+		/*!
+			\brief This function generates a Point.
+			Generated points depends on the world size which
+			means it cannot generate Points outside the current
+			"working" world.
+			\return return the generated point.
+		*/
+		Point randPoint();
+
+		/*!
+			\brief This function generates a Color.
+			\return the generated color.
+		*/
+		Color randColor();
+
+		/*!
+			\brief This function provide a way of generating random
+			complex and displayable hulls.
+			\param hullSize the size of the generated hull
+			(the number of Parts composing the hull).
+			\return the generated hull.
+		*/
+		PhysicalObject::Hull randHull(const int &hullSize);
+
+		/*!
+			\brief This function generates a basic rectangle part.
+			\return the generated part.
+		*/
+		PhysicalObject::Part randRectanglePart();
+
+		/*!
+			\brief This function generates a complex Part made of random
+			polygones.
+			\return the generated part.
+		*/
+		PhysicalObject::Part randComplexPart();
+
+		/*!
+			\brief This function generates random convex polygones.
+			\param polygonSize the number of points defining the polygone.
+			\return the generated polygone.
+		*/
+		Polygone randConvexPolygone(const int &polygoneSize = 3);
+
+		/*!
+			\brief Generate a random float within a specified range.
+			\param min the min possible value.
+			\param max the max possible value.
+			\return the generated float.
+		*/
+		float randFloat(const float& min, const float &max);
+
+		/*!
+			\brief Generate a random int within a specified range.
+			\param min the min possible value.
+			\param max the max possible value.
+			\return the generated int.
+		*/
+		int randInt(const int &min, const int &max);
+
+		/*!
+			\brief Generate a random float between 0 and 1
+			\return the generated float.
+		*/
+		float randColorFloat();
+
+		/*!
+			\brief Generate a random int between 0 and 1
+			\return the generated int.
+		*/
+		int randBool();
+	};
 }
-
 #endif
