+++ tests/testRandom.cpp
@@ -1,6 +1,5 @@
 #include <iostream>
 #include "WorldGenerator.h"
-#include "Randomizer2.h"
 #include <typeinfo>
 
 using namespace Enki;
